[NFC] Refactor in-process builtins into a new module
